+++ bin/string_utils.ml
@@ -17,11 +17,11 @@ let remove_leading_dots (s : string) =
     aux 0
   )
 
-let line_is_comment line =
-  CCString.starts_with ~prefix:"#" line
+let line_is_system_comment line =
+  CCString.starts_with ~prefix:";" line
 
-let line_is_blank_or_comment line =
-  line_is_comment line
+let line_is_blank_or_system_comment line =
+  line_is_system_comment line
   ||
   String.length (String.trim line) = 0